+++ src/anomalib/post_processing/__init__.py
@@ -5,7 +5,6 @@
 
 from .normalization import NormalizationMethod
 from .post_process import (
-    ThresholdMethod,
     add_anomalous_label,
     add_normal_label,
     anomaly_map_to_color_map,
@@ -23,5 +22,4 @@ __all__ = [
     "ImageResult",
     "NormalizationMethod",
     "Visualizer",
-    "ThresholdMethod",
 ]