+++ src/anomalib/utils/metrics/__init__.py
@@ -13,7 +13,6 @@ import torchmetrics
 from omegaconf import DictConfig, ListConfig
 
 from .anomaly_score_distribution import AnomalyScoreDistribution
-from .anomaly_score_threshold import AnomalyScoreThreshold
 from .aupr import AUPR
 from .aupro import AUPRO
 from .auroc import AUROC
@@ -21,8 +20,20 @@ from .collection import AnomalibMetricCollection
 from .min_max import MinMax
 from .optimal_f1 import OptimalF1
 from .pro import PRO
+from .thresholding import AdaptiveScoreThreshold, BaseAnomalyScoreThreshold, ManualThreshold
 
-__all__ = ["AUROC", "AUPR", "AUPRO", "OptimalF1", "AnomalyScoreThreshold", "AnomalyScoreDistribution", "MinMax", "PRO"]
+__all__ = [
+    "AUROC",
+    "AUPR",
+    "AUPRO",
+    "OptimalF1",
+    "AdaptiveScoreThreshold",
+    "AnomalyScoreDistribution",
+    "BaseAnomalyScoreThreshold",
+    "ManualThreshold",
+    "MinMax",
+    "PRO",
+]
 
 
 def metric_collection_from_names(metric_names: list[str], prefix: str | None) -> AnomalibMetricCollection: